changes with_span to use start/end span
previously dyalizer would not error on invalid contracts for functions
annotated with the decorator. presumably this was because the return
actually happens in a closure.

So, for example, the following code would pass dialyzer successfully

```elixir
@spec hello :: {:ok, :asdf}
@decorate with_span("hello")
def hello do
  :world
end
````

After this change it fails as expected

```
lib/spec_demo.ex:17:invalid_contract
The @spec for the function does not match the success typing of the function.

Function:
SpecDemo.hello/0

Success typing:
@spec hello() :: :world
```