addresses an issue setting error status on parent spans after exception
when two functions that are decorated with the with_span function are
nested and the child throws, the current span was not being set back to
the parent as expected. calling Tracer.set_status would attempt to set
the status on the child span which has been closed at that point and
fail meaning the parent's status would be undefined.